requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
Django==5.0
dnspython==2.4.2
h11==0.14.0
idna==3.6
motor==3.3.2
outcome==1.3.0.post0
pymongo==4.6.1
PySocks==1.7.1
python-dotenv==1.0.0
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
